Further work on shuffling around different states of the cart
