+++ plugins/woocommerce-blocks/assets/js/base/components/cart-checkout/shipping-calculator/index.tsx
@@ -1,3 +1,3 @@
 export * from './shipping-calculator';
-export * from './shipping-calculator-button';
+export * from './shipping-calculator-panel';
 export * from './context';
