+++ plugins/woocommerce-blocks/assets/js/base/components/cart-checkout/totals/shipping/index.tsx
@@ -131,6 +131,7 @@ export const TotalsShipping = ( {
 									}
 								/>
 								<ShippingAddress
+									showCalculator={ showCalculator }
 									shippingAddress={ shippingAddress }
 								/>
 							</>
