+++ plugins/woocommerce-blocks/assets/js/base/components/cart-checkout/totals/shipping/shipping-address.tsx
@@ -1,13 +1,11 @@
 /**
  * External dependencies
  */
-import { __, sprintf } from '@wordpress/i18n';
+import { __ } from '@wordpress/i18n';
 import { formatShippingAddress } from '@woocommerce/base-utils';
 import { ShippingAddress as ShippingAddressType } from '@woocommerce/settings';
 import {
-	ShippingLocation,
 	PickupLocation,
-	ShippingCalculatorButton,
 	ShippingCalculatorPanel,
 } from '@woocommerce/base-components/cart-checkout';
 import { CHECKOUT_STORE_KEY } from '@woocommerce/block-data';
@@ -15,10 +13,12 @@ import { useSelect } from '@wordpress/data';
 
 export interface ShippingAddressProps {
 	shippingAddress: ShippingAddressType;
+	showCalculator: boolean;
 }
 
 export const ShippingAddress = ( {
 	shippingAddress,
+	showCalculator,
 }: ShippingAddressProps ): JSX.Element | null => {
 	const prefersCollection = useSelect( ( select ) =>
 		select( CHECKOUT_STORE_KEY ).prefersCollection()
